Title: 최솟값 3개, Time: 1087ms, Memory: 0MB, Status: Time Limit Exceed - CodeTree

+++ 240118/최솟값 3개/top-3-smallest-number.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 int main() {
-	vector<int> v;
+	vector<long long> v;
 	long long n, input;
 	cin >> n;
